+++ src/blocks/accordion/editor.scss
