+++ src/blocks/accordion/index.js
@@ -1,0 +1,59 @@
+/**
+ * Registers a new block provided a unique name and an object defining its behavior.
+ *
+ * @see https://developer.wordpress.org/block-editor/reference-guides/block-api/block-registration/
+ */
+import { registerBlockType } from '@wordpress/blocks';
+
+/**
+ * Internal dependencies
+ */
+import Edit from './edit';
+import metadata from './block.json';
+import { InnerBlocks, useBlockProps } from '@wordpress/block-editor';
+
+/**
+ * Every block starts by registering a new block type definition.
+ *
+ * @see https://developer.wordpress.org/block-editor/reference-guides/block-api/block-registration/
+ */
+registerBlockType( metadata.name, {
+	/**
+	 * @see ./edit.js
+	 */
+	edit: Edit,
+	save: ( { attributes } ) => {
+		const blockProps = useBlockProps.save();
+		blockProps.className = `wmf-accordion-wrapper ${blockProps.className} ${attributes.fontColor}`;
+
+		return (
+			<div { ...blockProps } >
+				<InnerBlocks.Content />
+			</div>
+
+		);
+	},
+	deprecated: [
+        {
+            save: ( { attributes } ) => {
+				const blockProps = useBlockProps.save();
+				blockProps.className = `accordion-wrapper ${blockProps.className} ${attributes.fontColor}`;
+
+				return (
+					<div { ...blockProps } >
+						<InnerBlocks.Content />
+					</div>
+
+				);
+			},
+        },
+    ],
+} );
+
+// Block HMR boilerplate.
+if ( module.hot ) {
+	module.hot.accept();
+	const { deregister, refresh } = require( '../../helpers/hot-blocks.js' );
+	module.hot.dispose( deregister( metadata.name ) );
+	refresh( metadata.name, module.hot.data );
+}
