+++ src/frontend-global.scss
@@ -29,6 +29,7 @@
 @import "patterns/letter-from-the-ceo";
 @import "patterns/overlay";
 @import "patterns/welcome-page";
+@import "patterns/wmf-accordion";
 
 /* Endowment */
 @import "styles/endowment-overrides";
